+++ chromiumoxide_pdl/src/pdl/dep.rs
@@ -1,24 +1,25 @@
+use std::collections::HashSet;
+
 use once_cell::sync::Lazy;
-use std::collections::HashMap;
 
 // circularDeps is the list of types that can cause circular dependency
 // issues.
-static CIRCULAR_DEPS: Lazy<HashMap<&'static str, bool>> = Lazy::new(|| {
-    let mut m = HashMap::new();
-    m.insert("browser.browsercontextid", true);
-    m.insert("dom.backendnodeid", true);
-    m.insert("dom.backendnode", true);
-    m.insert("dom.nodeid", true);
-    m.insert("dom.node", true);
-    m.insert("dom.nodetype", true);
-    m.insert("dom.pseudotype", true);
-    m.insert("dom.rgba", true);
-    m.insert("dom.shadowroottype", true);
-    m.insert("network.loaderid", true);
-    m.insert("network.monotonictime", true);
-    m.insert("network.timesinceepoch", true);
-    m.insert("page.frameid", true);
-    m.insert("page.frame", true);
+static CIRCULAR_DEPS: Lazy<HashSet<&'static str>> = Lazy::new(|| {
+    let mut m = HashSet::new();
+    m.insert("browser.browsercontextid");
+    m.insert("dom.backendnodeid");
+    m.insert("dom.backendnode");
+    m.insert("dom.nodeid");
+    m.insert("dom.node");
+    m.insert("dom.nodetype");
+    m.insert("dom.pseudotype");
+    m.insert("dom.rgba");
+    m.insert("dom.shadowroottype");
+    m.insert("network.loaderid");
+    m.insert("network.monotonictime");
+    m.insert("network.timesinceepoch");
+    m.insert("page.frameid");
+    m.insert("page.frame");
     m
 });
 
@@ -27,6 +28,5 @@ static CIRCULAR_DEPS: Lazy<HashMap<&'static str, bool>> = Lazy::new(|| {
 pub(crate) fn is_circular_dep(dty: &str, ty_str: &str) -> bool {
     CIRCULAR_DEPS
         .get(format!("{}.{}", dty.to_lowercase(), ty_str.to_lowercase()).as_str())
-        .cloned()
-        .unwrap_or_default()
+        .is_some()
 }